+++ src/v2/services/keyvalue/index.ts
@@ -1,2 +1,7 @@
 export { KeyValueService } from "./keyvalue"
-export * from "./types"
+
+export type { KeyValueDisableArgs } from "./disable"
+export type { KeyValueGet, KeyValueGetArgs } from "./get"
+export type { KeyValueInfo } from "./info"
+export type { KeyValuePutArgs } from "./put"
+export type { KeyValueQuery, KeyValueQueryArgs } from "./query"
